LayoutTests/http/tests/security/model-element/model-element-crossorigin-blocked-with-use-credentials-and-all-access-header.html

<!DOCTYPE html> <!-- webkit-test-runner [ ModelElementEnabled=true ModelProcessEnabled=true ] -->
<html>
<head>
<title>&lt;model> blocked with use-credentials crossorigin and Access-Control-Allow-Origin allows all</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    if (window.testRunner)
        testRunner.waitUntilDone();

    window.addEventListener('load', async () => {
        const model = document.getElementById("blocked-model");
        assert_true(!!model, "Model element should exist");
        await model.ready.then(
            value => assert_unreached("Unexpected ready promise resolution."),
            reason => assert_true(reason.toString().includes("NetworkError"), "The ready promise is rejected with a NetworkError.")
        );
        testRunner.notifyDone();
    }, false);
</script>
</head>
<body>
    <div>Verify that the model is blocked even when the src allows access from all origins, if crossorigin is set to use-credentials</div>
    <model id='blocked-model' crossorigin='use-credentials'>
        <source src='http://localhost:8080/security/resources/model-access-control.py?file=cube.usdz&allowAll'/>
    </model>
</body>
</html>
